+++ packages/client/components/DashNavList/LeftNavHeader.tsx
@@ -5,5 +5,5 @@ interface Props {
 }
 export const LeftNavHeader = (props: Props) => {
   const {children} = props
-  return <div className='flex flex-col p-1 text-xs font-medium'>{children}</div>
+  return <div className='flex items-center p-1 text-xs font-medium'>{children}</div>
 }