+++ packages/server/dataloader/customLoaderMakers.ts
@@ -24,8 +24,6 @@ import getMeetingTaskEstimates, {
 import {
   selectMeetingSettings,
   selectNewMeetings,
-  selectPageAccess,
-  selectPageUserSortOrder,
   selectTasks,
   selectTeams
 } from '../postgres/select'
@@ -38,7 +36,7 @@ import {
   Team
 } from '../postgres/types'
 import {AnyMeeting, MeetingTypeEnum} from '../postgres/types/Meeting'
-import {Tierenum as TierEnum, type Pageroleenum} from '../postgres/types/pg'
+import {Tierenum as TierEnum} from '../postgres/types/pg'
 import {Logger} from '../utils/Logger'
 import getRedis from '../utils/getRedis'
 import isUserVerified from '../utils/isUserVerified'
@@ -1009,54 +1007,6 @@ export const highestTierForUserId = (parent: RootDataLoader) => {
     },
     {
       ...parent.dataLoaderOptions
-    }
-  )
-}
-
-export const pageAccessByUserId = (parent: RootDataLoader) => {
-  return new DataLoader<{pageId: number; userId: string}, Pageroleenum | null, string>(
-    async (keys) => {
-      const res = await selectPageAccess()
-        .where(({eb, refTuple, tuple}) =>
-          eb(
-            refTuple('pageId', 'userId'),
-            'in',
-            keys.map((key) => tuple(key.pageId, key.userId))
-          )
-        )
-        .execute()
-      return keys.map((key) => {
-        const rule = res.find(({pageId, userId}) => pageId === key.pageId && userId === key.userId)
-        return rule?.role ?? null
-      })
-    },
-    {
-      ...parent.dataLoaderOptions,
-      cacheKeyFn: (key) => `${key.pageId}:${key.userId}`
-    }
-  )
-}
-
-export const pageUserSortOrder = (parent: RootDataLoader) => {
-  return new DataLoader<{pageId: number; userId: string}, string | null, string>(
-    async (keys) => {
-      const res = await selectPageUserSortOrder()
-        .where(({eb, refTuple, tuple}) =>
-          eb(
-            refTuple('pageId', 'userId'),
-            'in',
-            keys.map((key) => tuple(key.pageId, key.userId))
-          )
-        )
-        .execute()
-      return keys.map((key) => {
-        const rule = res.find(({pageId, userId}) => pageId === key.pageId && userId === key.userId)
-        return rule?.sortOrder ?? null
-      })
-    },
-    {
-      ...parent.dataLoaderOptions,
-      cacheKeyFn: (key) => `${key.pageId}:${key.userId}`
     }
   )
 }