+++ packages/server/graphql/public/mutations/helpers/getPageNextSortOrder.ts
@@ -20,6 +20,7 @@ export const getPageNextSortOrder = async (
     .$if(!!parentPageId, (qb) => qb.where('parentPageId', '=', parentPageId!))
     .$if(!parentPageId, (qb) => qb.where('parentPageId', 'is', null))
     .where('PageAccess.userId', '=', viewerId)
+    // purposefully not excluding deleted docs because if they get restored then it could be a duplicate sortOrder
     .orderBy('sortOrder', 'desc')
     .limit(1)
     .executeTakeFirst()